Update Fourier series
In the first lab we will approximate some functions with sum of sine functions following
the Fourier series.
The functions you want to plot are:
1. The square wave:
build a 6 subplot figure to show the time domain, and use respectively 1, 3, 5, 10,
50, 500 frequencies. Each plot should have overlapped the real square function on a
different color. Remember to use a step of minimum 0.01 for the x axis. You can build
ab array t=0:0.01:1 (for example).
The Fourier series of the square wave is given by 4
π
∗
Pn
k=1,k=odd
sin(2πkt)
k
. Remember
that once you pass the array t to the sin function, this automatically becomes an array
of values.
In order to implement the SUM above you will need a for loop. Typically this is
done by initialising a variable (an array in this case of the same size as t) with zeros,
before you tart the for loop. Every time you go through the for loop you add the
newly caluclated sin() array to this sum value. i.e., sum=sum+sin(...);
Notice that the iteration is only done through the odd numbers, do your For loop will
need to start from 1 and advance at step of 2, for example ”for i=1:2:50”.
The square wave to be plotted in red is given by ’square(2 ∗ pi ∗ t)
0
. Again you
can pass the same array t as above.
IMPORTANT: you should first try to plot one Fourier approximation (for example
using the sum of 5 frequencies). Once this work you can then implement the outer for
loop to plot the 6 figures together.
Hint:
The following Matlab syntax might be useful. If you want to draw the 6 figures automatically using an outer For loop, you can iterate the loop over specific values (say:
1,3,5,10,50,500) in this way:
”for i=[1,3,5,10,50,500]
... code inside the loop that uses the variable i with the values in the bracket above...
end;”
IMPORTANT: You need to go through the Matlab tutorial first, so that you become
familiar with the Matlab syntax. You will find all the required Matlab commands in
the Matlab tutorial.
2. You should then draw the plot showing all sine components up to 10 sine functions
overlapping in a single plot. This will give a visual representation of all the 10 sine
functions at the same time. The example is shown below.
3. Draw the representation in the frequency domain of the square wave approximation
in exercise 1). For this you will need to do a stem plot, where each line represents the
amplitude of each sine wave on the sum. The plot will show all such amplitueds, in
sequence.