Update e Fast Fourier Transform
1 Introduction to the Fast Fourier Transform (FFT)
The Discrete Fourier Transform (DFT) is extremely important in the area
of spectrum analysis. It takes a discrete signal in the time domain and
transforms that signal into its discrete frequency domain representation;
this allows the processing of its frequency components through computerbased algorithms, which wouldn’t be able to work with a continuous analog
function.1
The Fast Fourier Transform (FFT) is just a clever implementation of the
DFT, specifically designed to be as fast as possible in terms of processing
time; its use has become so widespread that the acronyms FFT (a specific
implementation) and DFT (the generic analysis tool) have become interexchangeable.
1.1 Basic properties of the FFT
The time-continuous version of the DFT is periodic with period fs, where
fs is the sampling frequency of the discrete-time signal; it thus makes sense
to only define the DFT in the region between 0 and fs.
Furthermore, it’s easy to observe that the DFT is symmetric around
the centre point 0.5fs, which is called the Nyquist frequency (see Fig. ??).
For this reason, very often only half of the DFT is shown (e.g. the portion
between f = 0 and f = 0.5fs).
For a better understanding of these concepts, please refer to the lecture
slides "More on Spectrum".
1.2 FFT in Matlab
In Matlab, the FFT function is under the Signal Processing Toolbox. You
will need to install it (if you haven’t done so when you first installed Matlab).
1Note that this is different from the Discrete-Time Fourier Transform (DTFT), which
takes an discrete signal in the time domain and maps it into a continuous representation
in the frequency domain.
1
Figure 1: The DFT of a cosine with frequency equal to one tenth of the
sampling frequency.
The typical syntax for computing the FFT of a signal is FFT(x,N) where
x is the signal, x[n], you wish to transform, and N is the number of points
in the FFT. N must be at least as large as the number of samples in x[n].
2 Exercises
2.1 Exercise 1: Frequency Components With Noise
A common use of Fourier transforms is to find the frequency components of
a signal buried in a noisy time domain signal. The array of values given in
array.mat contains a signal made up of a number of sine waves with added
random noise.
• First load the variable array in matlab assigning it to the variable y:
use load(’array.mat’);
• Then plot the signal as it is (e.g., in the time domain) and check
whether you can identify the frequency of the sine functions making
up the signal (... probably not).
• Now do the FFT of the above signal, plot the frequency spectrum and
tell which frequencies is the signal made of.
2
You can use code similar to that in exercise 1 to carry out the FFT. Note
that the sampling frequency fs used to produce the array.mat file used was
1000 Hz. Use a value of at least N=1024 for the FFT.
2.2 Exercise 2: Spectrum of Music
In this exercise you will carry out an exercise similar to the one shown at
the end of the lecture slides "More on Spectrum".
From the file ’exercise notes.wav’, containing samples from two notes,
you will need to: separate the two notes into separate arrays, and find their
tune, by examining their spectrum of each array individually.
You will need to carry out the following steps in order to complete the
exercise:
• First load the sound samples using the following command: [notes,fsampling]
= audioread(’exercise notes.wav’);
• Then separate the two notes into two different arrays. In order to
understand where the second node begins you can plot the initial array
and/or listen to the two separate arrays that you create using the
function: sound(array). Notice that if you have an array ’test’ of 100
points you can divide it into two arrays of say 35 and 65 points doing:
test1=test(1:35); test2=test(36;100);
• Carry out the FFT for each array and plot the result. For the FFT
make sure you use a value of N equal or bigger than the number of
samples in the array you are calculating.
• Find the pitch of each of the two sounds by looking at the spectrum
and use the pdf document title ’frequency of musical notes.pdf’ to
identify the note.